refactor(utils): split utils.ts into domain-specific modules (git, dom, task, status)
