+++ src/components/features/chat/git-control-bar-push-button.tsx
@@ -1,6 +1,7 @@
 import { useTranslation } from "react-i18next";
 import ArrowUpIcon from "#/icons/u-arrow-up.svg?react";
-import { cn, getGitPushPrompt } from "#/utils/utils";
+import { cn } from "#/utils/utils";
+import { getGitPushPrompt } from "#/utils/git";
 import {
   gitControlBarActionButtonClassName,
   gitControlBarActionIconColor,
