+++ src/components/features/controls/agent-status.tsx
@@ -2,12 +2,12 @@ import { useTranslation } from "react-i18next";
 import { useEffect, useState } from "react";
 import { CircleCheck } from "lucide-react";
 import { useActiveConversation } from "#/hooks/query/use-active-conversation";
-import { getStatusCode } from "#/utils/status";
+import { getStatusCode, isTaskPolling } from "#/utils/status";
 import { ChatStopButton } from "../chat/chat-stop-button";
 import { AgentState } from "#/types/agent-state";
 import ClockIcon from "#/icons/u-clock-three.svg?react";
 import { ChatResumeAgentButton } from "../chat/chat-play-button";
-import { cn, isTaskPolling } from "#/utils/utils";
+import { cn } from "#/utils/utils";
 import { AgentLoading } from "./agent-loading";
 import { useConversationStore } from "#/stores/conversation-store";
 import CircleErrorIcon from "#/icons/circle-error.svg?react";
