+++ src/components/features/home/tasks/task-suggestions.tsx
@@ -3,7 +3,8 @@ import { useTranslation } from "react-i18next";
 import { TaskGroup } from "./task-group";
 import { useSuggestedTasks } from "#/hooks/query/use-suggested-tasks";
 import { TaskSuggestionsSkeleton } from "./task-suggestions-skeleton";
-import { cn, getDisplayedTaskGroups, getTotalTaskCount } from "#/utils/utils";
+import { cn } from "#/utils/utils";
+import { getDisplayedTaskGroups, getTotalTaskCount } from "#/utils/task";
 import { I18nKey } from "#/i18n/declaration";
 import { GitRepository } from "#/types/git";
 import { Typography } from "#/ui/typography";
